+++ app/cars/TrendChart.tsx
