+++ app/cars/fuel-types/[fuelType]/page.tsx
@@ -1,5 +1,5 @@
 import dynamic from "next/dynamic";
-import { fetchMonths } from "@/app/(default)/cars/utils/fetchMonths";
+import { fetchMonths } from "@/app/cars/utils/fetchMonths";
 import NoData from "@/components/NoData";
 import { StructuredData } from "@/components/StructuredData";
 import Typography from "@/components/Typography";
@@ -16,7 +16,7 @@ import { deslugify, slugify } from "@/utils/slugify";
 import type { Metadata } from "next";
 import type { WebPage, WithContext } from "schema-dts";
 
-type Params = Promise<{ vehicleType: string }>;
+type Params = Promise<{ fuelType: string }>;
 type SearchParams = Promise<{ [key: string]: string | string[] | undefined }>;
 
 const CarOverviewTrends = dynamic(
@@ -26,20 +26,22 @@ const MonthSelector = dynamic(() => import("@/components/MonthSelector"));
 
 export const generateMetadata = async (props: {
   params: Params;
+  searchParams: SearchParams;
 }): Promise<Metadata> => {
   const params = await props.params;
-  const { vehicleType } = params;
+  const { fuelType } = params;
 
-  const formattedVehicleType = deslugify(vehicleType);
-  const title = `${formattedVehicleType} Cars in Singapore`;
-  const description = `${formattedVehicleType} cars registrations by month. Explore registration trends, statistics and distribution by vehicle type for the month in Singapore.`;
-  // const images = `/api/og?title=Historical Trend&type=${vehicleType}`;
-  const canonical = `/cars/vehicle-types/${vehicleType}`;
+  const formattedFuelType = deslugify(fuelType);
+  const title = `${formattedFuelType} Cars in Singapore`;
+  const description = `${formattedFuelType} cars registrations by month. Explore registration trends, statistics and distribution by fuel type for the month in Singapore.`;
+  const canonical = `/cars/fuel-types/${fuelType}`;
 
   return {
     title,
     description,
     openGraph: {
+      title,
+      description,
       images: `${SITE_URL}/opengraph-image.png`,
       url: canonical,
       siteName: SITE_TITLE,
@@ -48,6 +50,8 @@ export const generateMetadata = async (props: {
     },
     twitter: {
       card: "summary_large_image",
+      title,
+      description,
       images: `${SITE_URL}/twitter-image.png`,
       site: "@sgcarstrends",
       creator: "@sgcarstrends",
@@ -58,50 +62,39 @@ export const generateMetadata = async (props: {
   };
 };
 
-const vehicleTypes = [
-  "hatchback",
-  "sedan",
-  "multi-purpose vehicle",
-  "station-wagon",
-  "sports utility vehicle",
-  "coupe/convertible",
-];
+const fuelTypes = ["petrol", "hybrid", "electric", "diesel"];
 
 export const generateStaticParams = () =>
-  vehicleTypes.map((vehicleType) => ({ vehicleType: slugify(vehicleType) }));
+  fuelTypes.map((fuelType) => ({ fuelType: slugify(fuelType) }));
 
-const CarsByVehicleTypePage = async (props: {
+const CarsByFuelTypePage = async (props: {
   params: Params;
   searchParams: SearchParams;
 }) => {
   const params = await props.params;
   const searchParams = await props.searchParams;
-  const { vehicleType } = params;
+  const { fuelType } = params;
 
   const [months, latestMonth]: [Month[], LatestMonth] = await fetchMonths();
 
   const month = searchParams?.month ?? latestMonth.cars;
   const cars = await fetchApi<Car[]>(
-    `${API_URL}/cars?vehicle_type=${vehicleType}&month=${month}`,
+    `${API_URL}/cars?fuel_type=${fuelType}&month=${month}`,
     { next: { tags: [RevalidateTags.Cars] } },
   );
 
-  if (cars.length === 0) {
-    return <NoData />;
-  }
-
   const filteredCars = mergeCarsByMake(cars);
 
-  const formattedVehicleType = deslugify(vehicleType);
+  const formattedFuelType = deslugify(fuelType);
 
-  const title = `${formattedVehicleType} Cars in Singapore`;
-  const description = `${formattedVehicleType} cars registrations by month. Explore registration trends, statistics and distribution by vehicle type for the month in Singapore.`;
+  const title = `${formattedFuelType} Cars in Singapore`;
+  const description = `${formattedFuelType} cars overview. Explore registration trends, statistics and distribution by fuel type for the month in Singapore.`;
   const structuredData: WithContext<WebPage> = {
     "@context": "https://schema.org",
     "@type": "WebPage",
     name: title,
     description,
-    url: `${SITE_URL}/cars/vehicle-types/${vehicleType}`,
+    url: `${SITE_URL}/cars/fuel-types/${fuelType}`,
     publisher: {
       "@type": "Organization",
       name: SITE_TITLE,
@@ -120,18 +113,17 @@ const CarsByVehicleTypePage = async (props: {
       <div className="flex flex-col gap-4">
         <div className="flex flex-col justify-between gap-2 xl:flex-row">
           <div className="flex items-start">
-            <Typography.H1>
-              {deslugify(vehicleType).toUpperCase()}
-            </Typography.H1>
+            <Typography.H1>{deslugify(fuelType).toUpperCase()}</Typography.H1>
           </div>
           <div className="items-end">
             <MonthSelector months={months} />
           </div>
         </div>
-        <CarOverviewTrends cars={filteredCars} />
+        {filteredCars.length > 0 && <CarOverviewTrends cars={filteredCars} />}
+        {filteredCars.length === 0 && <NoData />}
       </div>
     </>
   );
 };
 
-export default CarsByVehicleTypePage;
+export default CarsByFuelTypePage;
