+++ app/cars/makes/[make]/TrendChart.tsx
