+++ app/maintenance/page.tsx
@@ -1,9 +1,14 @@
 "use client";
 
 import { Suspense } from "react";
+import { MaintenanceNotice } from "@/app/maintenance/components/maintenance-notice";
 
 const MaintenancePage = () => {
-  return <Suspense fallback={null}></Suspense>;
+  return (
+    <Suspense fallback={null}>
+      <MaintenanceNotice />
+    </Suspense>
+  );
 };
 
 export default MaintenancePage;
